cli: split monolithic cli.rs into src/cli/ module tree
Reorganization-only change, no behavior or grammar change. The old
815-line src/cli.rs becomes:

  src/cli/mod.rs           — Cli, Command, run() dispatcher
  src/cli/exit.rs          — exit-code constants (0/1/2/3/4/5/130)
  src/cli/io.rs            — open_image_ro/rw helpers
  src/cli/parse.rs         — parse_size, pick_block_size,
                             split_mac_path, ZeroReader (+ unit tests)
  src/cli/api/mod.rs       — ApiGroup wrapper + dispatch
  src/cli/api/apm.rs       — api apm info
  src/cli/api/sgi.rs       — api sgi shrink
  src/cli/api/hfs.rs       — api hfs new/info/ls/put/get/rm/put-boot/
                             validate + resolve_hfs_offset/resolve_path

Sets up Phase A.3+ to layer in flat verbs and cross-cutting
infrastructure (--format, --log-level, IMG@N, …) without further
churn. Existing api-grammar tests pass unchanged.
